+++ datastruct/response.go
@@ -153,9 +153,9 @@ type (
 	}
 
 	HomeMerchantResponse struct {
+		MerchantID uint64        `gorm:"column:merchant_id" json:"merchant_id"`
 		Name       string        `gorm:"column:merchant_name" json:"merchant_name"`
 		AddressID  uint64        `gorm:"column:addresses_id" json:"-"`
-		MerchantID uint64        `gorm:"column:merchant_id" json:"-"`
 		Location   string        `json:"location"`
 		Product    []interface{} `json:"products"`
 	}