+++ CHANGELOG.md
@@ -89,3 +89,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add home partner app
 - Delete product
 - Product by id on mitra
+- Merchant and products